Users pay withdrawal fees
Withdrawals are now signed by the server but no transaction is sent by the
server. Instead, the signed withdrawal message is passed back to the client so
that a Web3 client like MetaMask can be used to send a transaction with this
signed message, allowing users to pay transaction fees.